Identity Util Methods moved to endpoint
