+++ components/configuration-mgt/org.wso2.carbon.identity.configuration.mgt.endpoint/src/main/java/org/wso2/carbon/identity/configuration/mgt/endpoint/impl/ResourceApiServiceImpl.java
@@ -29,7 +29,6 @@ import org.wso2.carbon.identity.configuration.mgt.endpoint.dto.AttributeDTO;
 import org.wso2.carbon.identity.configuration.mgt.endpoint.dto.ResourceAddDTO;
 import org.wso2.carbon.identity.configuration.mgt.endpoint.dto.ResourceFileDTO;
 import org.wso2.carbon.identity.configuration.mgt.endpoint.util.ConfigurationEndpointUtils;
-import org.wso2.carbon.identity.core.util.IdentityUtil;
 
 import java.io.InputStream;
 import java.net.URI;
@@ -38,8 +37,8 @@ import java.util.List;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
-import static org.wso2.carbon.identity.configuration.mgt.core.constant.ConfigurationConstants.FILE;
 import static org.wso2.carbon.identity.configuration.mgt.core.constant.ConfigurationConstants.RESOURCE_PATH;
+import static org.wso2.carbon.identity.configuration.mgt.endpoint.util.ConfigurationEndpointUtils.buildURIForHeader;
 import static org.wso2.carbon.identity.configuration.mgt.endpoint.util.ConfigurationEndpointUtils.getAttributeDTO;
 import static org.wso2.carbon.identity.configuration.mgt.endpoint.util.ConfigurationEndpointUtils.getAttributeFromDTO;
 import static org.wso2.carbon.identity.configuration.mgt.endpoint.util.ConfigurationEndpointUtils.getConfigurationManager;
@@ -189,7 +188,7 @@ public class ResourceApiServiceImpl extends ResourceApiService {
         try {
             ResourceFile resourceFile = getConfigurationManager()
                     .addFile(resourceType, resourceName, fileName, resourceFileInputStream);
-            return Response.created(IdentityUtil.buildURIForHeader(resourceFile.getPath())).build();
+            return Response.created(buildURIForHeader(resourceFile.getPath())).build();
         } catch (ConfigurationManagementClientException e) {
             return handleBadRequestResponse(e, LOG);
         } catch (ConfigurationManagementException e) {
